modify the price parameter for Payment.doPayment and ExternalAPI.doPayment from int to double modify the method for random to support double Add comments for PaymentRecordTest.java

+++ src/ExternalAPI/ExternalAPI.java
@@ -5,5 +5,5 @@ package ExternalAPI;
  * It is used to simulate the payment process of the external API
  */
 public interface ExternalAPI {
-    boolean doPayment(int price);
+    boolean doPayment(double price);
 }
