+++ src/Payment/Payment.java
@@ -11,7 +11,7 @@ public interface Payment {
      * @param price price of the product
      * @return true if the payment is successful, false otherwise
      */
-    boolean doPayment(int price);
+    boolean doPayment(double price);
 
     /**
      * Get the payment type
